+++ CHANGELOG.md
@@ -387,3 +387,4 @@
 - 2024-05-03: Impl mailbox, follow up, logo, mod, +2 query
 - 2024-05-03: Docs user, 0001 initial,  outreach base, bettercontact, +2
 - 2024-05-04: Valid mutate, follow up, user, descend, +4
+- 2024-05-05: Integ follow up, descend, logo, sender, +5